tools/voice-feedback/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "voice-feedback"
version = "0.1.0"
description = "Voice-based customer feedback analysis tool — transcribe, analyze, and feed insights into PostHog."
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Toolbox Team" }]
dependencies = [
    "faster-whisper>=1.0.0",
    "openai>=1.12.0",
    "posthog>=3.3.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
openai-whisper = ["openai-whisper>=20231117"]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "httpx>=0.27.0",
]

[project.scripts]
voice-feedback = "app.cli:cli"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
